+++ fdbclient/include/fdbclient/ManagementAPI.actor.h
@@ -34,7 +34,7 @@ standard API and some knowledge of the contents of the system key space.
 
 #include <string>
 #include <map>
-#include "fdbclient/BulkDumping.h"
+#include "fdbclient/BulkLoadAndDump.h"
 #include "fdbclient/GenericManagementAPI.actor.h"
 #include "fdbclient/NativeAPI.actor.h"
 #include "fdbclient/RangeLock.h"
