+++ fdbclient/include/fdbclient/StorageServerInterface.h
@@ -23,7 +23,7 @@
 #pragma once
 
 #include "fdbclient/Audit.h"
-#include "fdbclient/BulkDumping.h"
+#include "fdbclient/BulkLoadAndDump.h"
 #include "fdbclient/FDBTypes.h"
 #include "fdbclient/StorageCheckpoint.h"
 #include "fdbclient/StorageServerShard.h"
